+++ Fashionhero.Portal/Fashionhero.Portal.Shared.Model/Entity/Price.cs
@@ -16,10 +16,10 @@ namespace Fashionhero.Portal.Shared.Model.Entity
         public IProduct Product { get; set; }
 
         /// <inheritdoc />
-        public float NormalSell { get; set; }
+        public decimal NormalSell { get; set; }
 
         /// <inheritdoc />
-        public float? Discount { get; set; }
+        public decimal? Discount { get; set; }
 
         /// <inheritdoc />
         public CurrencyCode Currency { get; set; }
